requirements.txt
python-dotenv==1.0.1
openai==1.28.1
pyperclip==1.8.2
pygame==2.5.2